+++ src/app/products/components/product-details/product-details.component.ts
@@ -2,6 +2,7 @@ import { Component, inject, OnInit } from "@angular/core";
 import { ActivatedRoute, ParamMap, Route, Router } from "@angular/router";
 import { map, Observable, switchMap } from "rxjs";
 
+import { CartService } from "../../../cart/services/cart.service";
 import { Product } from "../../models/product";
 import { ProductsService } from "../../services/products.service";
 
@@ -14,6 +15,7 @@ export class ProductDetailsComponent implements OnInit {
 
     product$!: Observable<Product>;
 
+    private cartService: CartService = inject(CartService);
     private service: ProductsService = inject(ProductsService);
     private route: ActivatedRoute = inject(ActivatedRoute);
 
@@ -23,5 +25,14 @@ export class ProductDetailsComponent implements OnInit {
             this.service.getProducts().pipe(
                 map((products: Product[]) => products.find(item => item.id === params.get('id'))!)))
         );
-      }
+    }
+
+    onAddProduct(item: Product) {
+      this.cartService.addItem({
+        id: item.id,
+        name: item.name,
+        price: item.price,
+        amount: 1
+      })
+    }
 }